Renamed ioStream_loopback to ioStream_pipe (which is actually the correct terminology based upon its functionality). Implemented a true ioStream_loopback
