+++ include/arch-common/cxa_ioStream_pipe.h
@@ -13,8 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifndef CXA_IOSTREAM_LOOPBACK_H_
-#define CXA_IOSTREAM_LOOPBACK_H_
+#ifndef CXA_IOSTREAM_PIPE_H_
+#define CXA_IOSTREAM_PIPE_H_
 
 
 /**
@@ -32,29 +32,36 @@
 
 
 // ******** global macro definitions ********
-#ifndef CXA_IOSTREAM_LOOPBACK_BUFFER_SIZE_BYTES
-	#define CXA_IOSTREAM_LOOPBACK_BUFFER_SIZE_BYTES				128
+#ifndef CXA_IOSTREAM_PIPE_BUFFER_SIZE_BYTES
+	#define CXA_IOSTREAM_PIPE_BUFFER_SIZE_BYTES				128
 #endif
 
 
 // ******** global type definitions *********
 /**
  * @public
- * @brief "Forward" declaration of the cxa_ioStream_loopback_t object
+ * @brief "Forward" declaration of the cxa_ioStream_pipe_t object
  */
-typedef struct cxa_ioStream_loopback cxa_ioStream_loopback_t;
+typedef struct cxa_ioStream_pipe cxa_ioStream_pipe_t;
 
 
-struct cxa_ioStream_loopback
+struct cxa_ioStream_pipe
 {
-	cxa_ioStream_t super;
+	cxa_ioStream_t endPoint1;
+	cxa_ioStream_t endPoint2;
 
-	cxa_fixedFifo_t fifo;
-	uint8_t fifo_raw[CXA_IOSTREAM_LOOPBACK_BUFFER_SIZE_BYTES];
+	cxa_fixedFifo_t fifo_ep1Read;
+	uint8_t fifo_ep1Read_raw[CXA_IOSTREAM_PIPE_BUFFER_SIZE_BYTES];
+
+	cxa_fixedFifo_t fifo_ep2Read;
+	uint8_t fifo_ep2Read_raw[CXA_IOSTREAM_PIPE_BUFFER_SIZE_BYTES];
 };
 
 
 // ******** global function prototypes ********
-void cxa_ioStream_loopback_init(cxa_ioStream_loopback_t *const ioStreamIn);
+void cxa_ioStream_pipe_init(cxa_ioStream_pipe_t *const ioStreamIn);
 
-#endif // CXA_IOSTREAM_LOOPBACK_H_
+cxa_ioStream_t* cxa_ioStream_pipe_getEndpoint1(cxa_ioStream_pipe_t* const ioStreamIn);
+cxa_ioStream_t* cxa_ioStream_pipe_getEndpoint2(cxa_ioStream_pipe_t* const ioStreamIn);
+
+#endif // CXA_IOSTREAM_PIPE_H_
